Make unique ids non-strict, remove an existential type
These fields were unpacked for space efficiency before, but
some call sites had to re-box the ids anyway and this was
causing excessive allocations on each uid access.  They are now
not unboxed and the allocations are gone.

This change also eliminates the existential type for Value, which is
now ValueContent.  This is a bit simpler and requires one fewer type
extension.